Move container_exec to components
