+++ app/grandchallenge/components/signals.py
@@ -1,11 +1,11 @@
 from django.db.models.signals import post_save
 from django.dispatch import receiver
 
-from grandchallenge.container_exec.models import (
+from grandchallenge.components.models import (
     ContainerExecJobModel,
     ContainerImageModel,
 )
-from grandchallenge.container_exec.tasks import validate_docker_image_async
+from grandchallenge.components.tasks import validate_docker_image_async
 from grandchallenge.core.utils import disable_for_loaddata
