+++ app/grandchallenge/components/tasks.py
@@ -13,10 +13,8 @@ from django.db import OperationalError
 from django.db.models import DateTimeField, ExpressionWrapper, F
 from django.utils.timezone import now
 
-from grandchallenge.container_exec.backends.docker import (
-    ContainerExecException,
-)
-from grandchallenge.container_exec.emails import send_invalid_dockerfile_email
+from grandchallenge.components.backends.docker import ContainerExecException
+from grandchallenge.components.emails import send_invalid_dockerfile_email
 from grandchallenge.jqfileupload.widgets.uploader import StagedAjaxFile
