--- app/grandchallenge/container_exec/__init__.py
@@ -1,1 +1,0 @@
-default_app_config = "grandchallenge.container_exec.apps.CoreConfig"
