+++ app/tests/container_exec_tests/test_tasks.py
@@ -3,7 +3,7 @@ from datetime import datetime, timedelta
 import pytest
 
 from grandchallenge.algorithms.models import Job as AlgorithmJob
-from grandchallenge.container_exec.tasks import mark_long_running_jobs_failed
+from grandchallenge.components.tasks import mark_long_running_jobs_failed
 from grandchallenge.evaluation.models import Job as EvaluationJob
 from tests.algorithms_tests.factories import AlgorithmJobFactory
 from tests.factories import JobFactory as EvaluationJobFactory
