+++ app/tests/evaluation_tests/test_tasks.py
@@ -4,7 +4,7 @@ import docker
 import pytest
 from django.core.exceptions import ValidationError
 
-from grandchallenge.container_exec.tasks import validate_docker_image_async
+from grandchallenge.components.tasks import validate_docker_image_async
 from grandchallenge.evaluation.models import Method
 from tests.factories import MethodFactory, SubmissionFactory
