+++ app/tests/workstations_tests/test_models.py
@@ -5,7 +5,7 @@ from django.core.exceptions import ObjectDoesNotExist
 from docker.errors import NotFound
 from rest_framework.authtoken.models import Token
 
-from grandchallenge.container_exec.tasks import stop_expired_services
+from grandchallenge.components.tasks import stop_expired_services
 from grandchallenge.workstations.models import Session, Workstation
 from tests.factories import (
     SessionFactory,
